libopensdg: Fix busylooping
Fix stupid bug with poll() timeout always being calculated as 0

+++ library/utils.h
@@ -108,7 +108,7 @@ static unsigned long long timestamp(void)
     struct timespec ts;
 
     clock_gettime(CLOCK_MONOTONIC, &ts);
-    return ts.tv_sec * 1000 + ts.tv_nsec / 1000000;
+    return (unsigned long long)ts.tv_sec * 1000 + ts.tv_nsec / 1000000;
 }
 
 #endif
